+++ js/lib/intro.js
@@ -1,6 +1,10 @@
 import getTemplate from './get-template';
-import getPage from './get-page';
+import transitionTo from './utilities';
 import greetingElement from './greeting';
+
+let intro = {
+  'description': '<sup>*</sup>Это не фото. Это рисунок маслом нидерландского художника-фотореалиста Tjalf Sparnaay.'
+};
 
 /**
  * @type {string}
@@ -8,8 +12,7 @@ import greetingElement from './greeting';
 const template =
   `<div id="intro" class="intro">
     <h1 class="intro__asterisk">*</h1>
-    <p class="intro__motto"><sup>*</sup> Это не фото. Это рисунок маслом нидерландского художника-фотореалиста Tjalf
-        Sparnaay.</p>
+    <p class="intro__motto">${intro.description}</p>
   </div>`;
 
 /**
@@ -18,10 +21,6 @@ const template =
 const introElement = getTemplate(template);
 const startBtn = introElement.querySelector('.intro__asterisk');
 
-startBtn.onclick = (e) => {
-  e.preventDefault();
-
-  getPage(greetingElement);
-};
+startBtn.addEventListener('click', transitionTo(greetingElement));
 
 export default introElement;
