+++ js/lib/stats.js
